Split up CLI and config tests
